Revert "Connor updates"
