--- services/auth_service/config/auth_config.yaml
@@ -1,10 +1,0 @@
-database:
-  uri: "sqlite:///auth.db"
-  track_modifications: false
-
-profile-service:
-  url: "localhost:5001"
-
-flask:
-  secret_key: 'ECE4150-Lab4'
-
