--- services/auth_service/model.py
@@ -1,5 +1,0 @@
-from database import db
-
-class Auth(db.Model):
-    profile_id = db.Column(db.Integer, primary_key=True)
-    access_token = db.Column(db.String(500), nullable=False)
