--- services/chatroom_service/config/chatroom_config.yaml
@@ -1,10 +1,0 @@
-database:
-  uri: "sqlite:///chatrooms.db"
-  track_modifications: false
-
-profile-service:
-  url: "localhost:5001"
-
-flask:
-  secret_key: 'ECE4150-Lab4'
-
