--- services/chatroom_service/database.py
@@ -1,3 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
