--- services/login_service/config/login_config.yaml
@@ -1,10 +1,0 @@
-database:
-  uri: ""
-  track_modifications: false
-
-profile-service:
-  url: "localhost:5001"
-
-flask:
-  secret_key: 'ECE4150-Lab4'
-
