+++ templates/chatrooms.html
@@ -2,41 +2,22 @@
 
 {% block content %}
 <h1>Chatrooms</h1>
-<!-- <a href="{{ url_for('create_chatroom') }}" class="create-chatroom-button">+ Create New Chatroom</a> -->
-<!-- CHANGE HREF SO THEY KEEP THE PROFILE ID AND ACCESS TOKEN ACROSS CHANGED -->
-<a href="{{ url_for('create_chatroom', profile_id=request.args.get('profile_id'), access_token=request.args.get('access_token')) }}" class="create-chatroom-button">+ Create New Chatroom</a>
-
-<!-- Success Message -->
-<div id="success-message" class="success-popup">
-  <p>You have successfully joined the chatroom!</p>
-  <a href="#" class="close-popup">Close</a>
-</div>
-
-<!-- Left Chatroom Message -->
-<div id="left-message" class="left-popup">
-  <p>You have successfully left the chatroom!</p>
-  <a href="#" class="close-left-popup">Close</a>
-</div>
-
-<!-- <p>Current User: {{ current_user }}</p> -->
-<ul class="chatroom-list">
+<a href="{{ url_for('create_chatroom') }}">+ Create New Chatroom</a>
+<ul>
   {% for room in rooms %}
-  <li class="chatroom-item">
-    <div class="chatroom-info">
-      <a href="{{ url_for('chatroom', room_id=room.chatroom_id, profile_id=request.args.get('profile_id'), access_token=request.args.get('access_token')) }}" class="chatroom-name">{{ room.chatroom_name }}</a>
-    </div>
-    {% set member_ids = chatroomMembers | selectattr('chatroom_id', 'equalto', room.chatroom_id) | map(attribute='profile_id') | list %}
-    <!-- <p>Member IDs for this chatroom: {{ member_ids }}</p> -->
-    {% if current_user | int in member_ids %}
+  <li>
+      <a href="/chatroom/{{ room.chatroom_id }}">{{ room.chatroom_name }}</a>
+      {% set member_ids = room.chatroom_members | map(attribute='profile_id') | list %}
+      {% if session['user_id'] in member_ids %}
           <!-- User is already a member -->
-          <form action="{{ url_for('leave_chatroom', room_id=room.chatroom_id, profile_id=request.args.get('profile_id'),access_token=request.args.get('access_token')) }}#left-message" method="post" style="display:inline;">
-              <button type="submit" class="leave-button">Leave</button>
+          <form action="{{ url_for('leave_chatroom', room_id=room.chatroom_id) }}" method="post" style="display:inline;">
+              <button type="submit">Leave</button>
           </form>
       {% else %}
           <!-- User is not a member -->
-        <form action="{{ url_for('join_chatroom', room_id=room.chatroom_id, profile_id=request.args.get('profile_id'),access_token=request.args.get('access_token')) }}#success-message" method="post" style="display:inline;">
-            <button type="submit" class="join-button">Join</button>
-        </form>
+          <form action="{{ url_for('join_chatroom', room_id=room.chatroom_id) }}" method="post" style="display:inline;">
+              <button type="submit">Join</button>
+          </form>
       {% endif %}
   </li>
 {% endfor %}
