+++ templates/login.html
@@ -1,18 +1,10 @@
 {% extends "base.html" %}
 {% block content %}
-<h1>Chatroom Login</h1>
-<form method="post" class="basic-form">
-  <div class="form-group">
-  <label for="username">Username: </label>
-  <input type="text" name="username" class="form-input">
-  </div>
-  <div class="form-group">
-    <label for="password">Password: </label>
-    <input type="password" name="password" class="form-input">
-  </div>
-  
-  <button type="submit" id="loginbutton" class="button">Login</button> <br><br>
-  <a href="/register" class="register-link">Register here</a>
+<h1>Login</h1>
+<form method="post">
+  Username: <input type="text" name="username"><br>
+  Password: <input type="password" name="password"><br>
+  <button type="submit">Login</button>
 </form>
-
+<a href="/register">Register here</a>
 {% endblock %}
