+++ templates/register.html
@@ -1,19 +1,10 @@
 {% extends "base.html" %}
 {% block content %}
 <h1>Register</h1>
-<form method="post" class="login-form">
-  <div class="form-group">
-    <label for="username">Username: </label>
-    <input type="text" name="username" class="form-input">
-  </div>
-  <div class="form-group">
-    <label for="password">Password: </label>
-    <input type="text" name="password" class="form-input">
-  </div>
-  <div class="form-group">
-    <label for="password">Bio: </label>
-    <textarea name="bio" class="form-input"></textarea><br>
-  </div>
-  <button type="submit" class="button">Register</button>
+<form method="post">
+  Username: <input type="text" name="username"><br>
+  Password: <input type="password" name="password"><br>
+  Bio: <textarea name="bio"></textarea><br>
+  <button type="submit">Register</button>
 </form>
 {% endblock %}
